# backend/src/API.yml
openapi: 3.0.3
info:
  description: This is the interactive documentation for the Innovatiehuis Intranet RESTful API
  version: "0.0.0"
  title: Innovatieplatform REST API
servers:
  - url: http://127.0.0.1:5000/api



  # Paths supported by the server application
paths:
...
  /projects/{project_id}:
    get:
      operationId: src.endpoints.projects.read_one
      tags:
        - Projects
      summary: Read one project
      parameters:
        - $ref: '#/components/parameters/project_id'
      responses:
        200:
          description: Successfully read project operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Project'
    put:
      operationId: src.endpoints.projects.update
      tags:
        - Projects
      summary: Update a value of project
      parameters:
        - $ref: '#/components/parameters/project_id'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                project_name:
                  type: string
                description:
                  type: string

      responses:
        200:
          description: Project updated
    delete:
      operationId: src.endpoints.projects.delete
      tags:
        - Projects
      summary: Deletes the specified project
      parameters:
        - $ref: '#/components/parameters/project_id'
      responses:
        200:
          description: Project deleted
...
  /announcements:
    get:
      operationId: src.endpoints.announcements.read_global
      tags:
        - Global Announcements
      summary: Reads all global announcements (those not tied to any project)
      responses:
        200:
          description: Successfully retreieved global announcements
    post:
      operationId: src.endpoints.announcements.create_global
      tags:
        - Global Announcements
      summary: Create a new global announcement (not tied to any project)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Announcement'
      responses:
        200:
          description: Successfully posted announcement to project
  /announcements/{announcement_id}:
    get:
      operationId: src.endpoints.announcements.read_one
      tags:
        - Announcements
      summary: Reads a specific announcement
      parameters:
        - $ref: '#/components/parameters/announcement_id'
      responses:
        200:
          description: Successfully retrieved single announcement
    put:
      operationId: src.endpoints.announcements.update
      tags:
        - Announcements
      summary: Update title and/or content of a specific announcement
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Announcement'
      parameters:
        - $ref: '#/components/parameters/announcement_id'
      responses:
        200:
          description: Successfully posted announcement to project
    delete:
      operationId: src.endpoints.announcements.delete
      tags:
        - Announcements
      summary: Deletes specific announcement
      parameters:
        - $ref: '#/components/parameters/announcement_id'
      responses:
        200:
          description: Successfully removed announcement
  /announcements/{announcement_id}/replies:
    get:
      operationId: src.endpoints.announcements.read_replies
      tags:
        - Announcements
        - Replies
      summary: Get all replies to an announcement
      parameters:
        - $ref: '#/components/parameters/announcement_id'
      responses:
        200:
          description: Successfully retrieved replies to announcements
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
    post:
      operationId: src.endpoints.announcements.add_reply
      tags:
        - Announcements
        - Replies
      summary: Post reply to an announcement
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                content:
                  type: string
                title:
                  type: string
      parameters:
        - $ref: '#/components/parameters/announcement_id'
      responses:
        200:
          description: Successfully posted reply to announcement


# =========================================REPLIES===============================================
...
  /users/{user_id}:
    get:
      operationId: src.endpoints.users.read_one
      tags:
        - Users
      summary: Get specified user
      parameters:
        - $ref: '#/components/parameters/user_id'
      responses:
        200:
          description: Successfully read user
    put:
      operationId: src.endpoints.users.update
      tags:
        - Users
      summary: Edit user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/User' # BUT NOT PASSWORD_HASH
      parameters:
        - $ref: '#/components/parameters/user_id'
      responses:
        200:
          description: Successfully edited user
        403:
          description: Request denied due to lack of permissions
    delete:
      operationId: src.endpoints.users.delete
      tags:
        - Users
      summary: Delete specified user
      parameters:
        - $ref: '#/components/parameters/user_id'
      responses:
        200:
          description: Successfully deleted user
...
  /auth/password:
    put:
      operationId: src.endpoints.auth.change_password
      tags:
        - Authentication
      summary: Request JSON Web Token
      parameters:
        - in: query
          name: resettoken
          schema:
            type: string
          required: false
      responses:
        200:
          description: User is authenticated and receives a JSON Web Token
          content:
            'application/json':
              schema:
                type: string
        401:
          description: Unauthorized
...
components:
  parameters:
    project_id:
      in: path
      name: project_id
      required: true
      schema:
        type: string
      description: The ID number of a project
...
    user_id:
      in: path
      name: user_id
      required: true
      schema:
        type: string
      description: The ID number of a user
    announcement_id:
      in: path
      name: announcement_id
      required: true
      schema:
        type: string
      description: The ID number of an announcement
...
  schemas:
    Project:
      type: object
      properties:
        projectid:
          type: integer
        project_name:
          type: string
        is_archived:
          type: boolean
        create:
          type: string
        last_updated:
          type: string
    User:
      type: object
      properties:
        first_name:
          type: string
        last_name:
          type: string
        email:
          type: string
        phone_number:
          type: string
        roleid:
          type: integer
        screening_status:
          type: integer
        password_hash:
          type: string
    Announcement:
      type: object
      properties:
        content:
          type: string
        title:
          type: string
...